Move more files used by media only.
