+++ assets/src/edit-story/elements/media/visibleImage.js
